requirements.txt
kagglehub==0.3.10
opencv-python==4.11.0.86
pillow==11.1.0
sckit-learn==1.6.1
tqdm==4.67.1
torch==2.6.0
transformers==4.49.0